+++ EquiRectangular/src/DPC++-TwoDevices/TwoDevices/TimingStats.hpp
@@ -43,19 +43,19 @@ private:
 
 	// m_durationWarmup holds the very first duration for each timing type.  The "warmup" run
 	// is often much longer than the subsequent runs since the kernel may need to be compiled, etc.
-	std::chrono::duration<double> m_durationWarmup[TIMING_MAX][ALL_STATS];
+	std::chrono::duration<float> m_durationWarmup[TIMING_MAX][ALL_STATS];
 	int m_warmupIterations[TIMING_MAX][ALL_STATS];
 	// m_iterations holds the number of times a given timing type has been reported (minus the warmup run)
 	int m_iterations[TIMING_MAX][ALL_STATS];
 	// m_durationsSum holds the sum of all the post-warmup runs
-	std::chrono::duration<double> m_durationsSum[TIMING_MAX][ALL_STATS];
+	std::chrono::duration<float> m_durationsSum[TIMING_MAX][ALL_STATS];
 	// m_lapIterations holds a "lap" counter that helps when looking at instantaneous runs rather than the total
 	// run.  For example if the algorithm takes a lot of time to compute when parameters are changed, but less
 	// time when flipping frames, then the lap time can be used to show the latest frame while the m_iterations
 	// and m_durationsSum will report the overall results including both the runs where parameters were changed
 	// and those where they were not.
 	int m_lapIterations[TIMING_MAX][ALL_STATS];
-	std::chrono::duration<double> m_lapDurationsSum[TIMING_MAX][ALL_STATS];
+	std::chrono::duration<float> m_lapDurationsSum[TIMING_MAX][ALL_STATS];
 	std::chrono::high_resolution_clock::time_point m_startWarmupTime;
 	std::chrono::high_resolution_clock::time_point m_startNonWarmupTime;
 	bool m_bFirstNonWarmup;
